emutet/requirements.txt
requests
protobuf
python-magic
